Restructure UI towards what I'm imagining
A composer will allow you to write in basic Markdown (for now) an entry
about the photo shown on the left. Above the composer are other photos
that were taken around the same time period as the primary one. You will
be able to select those to add to your entry as well.